# backend/swagger_output.json
{
  "swagger": "2.0",
  "info": {
    "version": "1.0.0",
    "title": "PortFeed",
    "description": "API PORTFEED, rede social de portfólios."
  },
  "host": "localhost:3000",
  "basePath": "/",
  "tags": [
    {
      "name": "Projeto",
      "description": "Rota para entidade projeto. Cuida dos projetos de um user"
    },
    {
      "name": "Postagem",
      "description": "Rota para entidade postagem. Cuida dos posts que um user faz"
    },
    {
      "name": "Usuario",
      "description": "Rota para entidade usuario. Cuida do perfil"
    }
  ],
  "schemes": [
    "http",
    "https"
  ],
  "securityDefinitions": {
    "bearerAuth": {
      "type": "apiKey",
      "name": "Authorization",
      "scheme": "bearer",
      "in": "header"
    }
  },
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "paths": {
    "/users/": {
      "get": {
        "tags": [],
        "description": "",
        "produces": [
          "application/json"
        ],
        "parameters": [],
        "responses": {}
      }
    },
    "/users/signup": {
      "post": {
        "tags": [
          "Users"
        ],
        "description": "Registra usuário",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "obj",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "username": {
                  "example": "any"
                },
                "password": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "schema": {
              "$ref": "#/definitions/SignUp"
            },
            "description": "Registrado com sucesso."
          },
          "500": {
            "schema": {
              "$ref": "#/definitions/SignUpErro"
            },
            "description": "O usuário já está cadastrado."
          }
        }
      }
    },
    "/users/login": {
      "post": {
        "tags": [
          "Users"
        ],
        "description": "Loga usuário",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "obj",
            "in": "body",
            "type": "object",
            "schema": {
              "$ref": "#/definitions/Login"
            }
          }
        ],
        "responses": {
          "200": {
            "schema": {
              "$ref": "#/definitions/LoginResposta"
            },
            "description": "Logado com sucesso."
          },
          "401": {
            "description": "Erro na autenticação."
          }
        }
      }
    },
    "/users/logout": {
      "get": {
        "tags": [
          "Users"
        ],
        "description": "Desloga usuário",
        "parameters": [
          {
            "name": "obj",
            "in": "body",
            "type": "object",
            "schema": {
              "$ref": "#/definitions/Logout"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Deslogado com sucesso."
          }
        }
      }
    },
    "/projetos/": {
      "get": {
        "tags": [
          "Projeto"
        ],
        "description": "Endpoint para encontrar os projetos.",
        "produces": [
          "application/json"
        ],
        "parameters": [],
        "responses": {
          "200": {
            "schema": {
              "$ref": "#/definitions/Projetos"
            },
            "description": "Projeto encontrado."
          },
          "401": {
            "description": "Acesso não autorizado, necessário fazer login"
          },
          "404": {
            "description": "Projeto não encontrado."
          },
          "500": {
            "description": "Server ou banco fora do ar."
          }
        }
      },
      "post": {
        "tags": [
          "Projetos"
        ],
        "description": "Endpoint para criar um novo projeto.",
        "produces": [
          "application/json"
        ],
        "parameters": [],
        "responses": {
          "200": {
            "description": "Projeto criado."
          },
          "401": {
            "description": " Acesso não autorizado, necessário fazer login"
          },
          "500": {
            "description": "Server ou banco fora do ar."
          }
        }
      },
      "options": {
        "tags": [],
        "description": "",
        "parameters": [],
        "responses": {}
      }
    },
    "/projetos/{id}": {
      "get": {
        "tags": [
          "Projetos"
        ],
        "description": "Endpoint que encontra um projeto através de seu id.",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string",
            "description": "ID do projeto."
          }
        ],
        "responses": {
          "200": {
            "schema": {
              "$ref": "#/definitions/Projeto"
            },
            "description": "Projeto existe e é encontrado."
          },
          "401": {
            "description": " Acesso não autorizado, necessário fazer login"
          },
          "404": {
            "description": "Projeto não encontrado"
          },
          "500": {
            "description": "Server ou banco fora do ar."
          }
        }
      },
      "put": {
        "tags": [
          "Projeto"
        ],
        "description": "Endpoint para encontrar e atualizar um projeto através do id.",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string",
            "description": "ID do projeto."
          }
        ],
        "responses": {
          "200": {
            "schema": {
              "$ref": "#/definitions/Projeto"
            },
            "description": "Projeto foi editado."
          },
          "401": {
            "description": " Acesso não autorizado, necessário fazer login"
          },
          "404": {
            "description": "Projeto não encontrado"
          },
          "500": {
            "description": "Server ou banco fora do ar."
          }
        }
      },
      "delete": {
        "tags": [
          "Projeto"
        ],
        "description": "Endpoint para encontrar e excluir um projeto através do id.",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string",
            "description": "ID do projeto"
          }
        ],
        "responses": {
          "200": {
            "description": "Projeto apagado"
          },
          "401": {
            "description": " Acesso não autorizado, necessário fazer login"
          },
          "404": {
            "description": "Projeto não encontrado"
          },
          "500": {
            "description": "Server ou banco fora do ar."
          }
        }
      }
    },
    "/postagem/": {
      "get": {
        "tags": [
          "Postagem"
        ],
        "description": "Endpoint para encontrar as postagens.",
        "produces": [
          "application/json"
        ],
        "parameters": [],
        "responses": {
          "200": {
            "schema": {
              "$ref": "#/definitions/Postagem"
            },
            "description": "Post encontrado."
          },
          "401": {
            "description": "Acesso não autorizado, necessário fazer login"
          },
          "404": {
            "description": "Post não encontrado."
          },
          "500": {
            "description": "Server ou banco fora do ar."
          }
        }
      },
      "post": {
        "tags": [
          "Postagem"
        ],
        "description": "Endpoint para criar uma nova postagem.",
        "produces": [
          "application/json"
        ],
        "parameters": [],
        "responses": {
          "200": {
            "description": "Postagem criada."
          },
          "401": {
            "description": " Acesso não autorizado, necessário fazer login"
          },
          "500": {
            "description": "Server ou banco fora do ar."
          }
        }
      },
      "options": {
        "tags": [],
        "description": "",
        "parameters": [],
        "responses": {}
      }
    },
    "/postagem/{id}": {
      "get": {
        "tags": [
          "Postagem"
        ],
        "description": "Endpoint que encontra uma postagem através de seu id.",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string",
            "description": "ID da postagem."
          }
        ],
        "responses": {
          "200": {
            "schema": {
              "$ref": "#/definitions/Postagem"
            },
            "description": "Postagem existe e é encontrada."
          },
          "401": {
            "description": " Acesso não autorizado, necessário fazer login"
          },
          "404": {
            "description": "Post não encontrado."
          },
          "500": {
            "description": "Server ou banco fora do ar."
          }
        }
      },
      "put": {
        "tags": [
          "Postagem"
        ],
        "description": "Endpoint para encontrar e atualizar uma postagem através do id.",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string",
            "description": "ID da postagem."
          }
        ],
        "responses": {
          "200": {
            "schema": {
              "$ref": "#/definitions/Postagem"
            },
            "description": "Postagem foi editada."
          },
          "401": {
            "description": " Acesso não autorizado, necessário fazer login"
          },
          "404": {
            "description": "Post não encontrado."
          },
          "500": {
            "description": "Server ou banco fora do ar."
          }
        }
      },
      "delete": {
        "tags": [
          "Postagem"
        ],
        "description": "Endpoint para encontrar e excluir uma postagem através do id.",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string",
            "description": "ID da postagem"
          }
        ],
        "responses": {
          "200": {
            "description": "Postagem apagada"
          },
          "401": {
            "description": " Acesso não autorizado, necessário fazer login"
          },
          "404": {
            "description": "Post não encontrado."
          },
          "500": {
            "description": "Server ou banco fora do ar."
          }
        }
      }
    },
    "/usuario/": {
      "get": {
        "tags": [
          "Usuário"
        ],
        "description": "Endpoint para encontrar os usuários.",
        "produces": [
          "application/json"
        ],
        "parameters": [],
        "responses": {
          "200": {
            "schema": {
              "$ref": "#/definitions/Usuarios"
            },
            "description": "Usuários encontrados."
          },
          "401": {
            "description": "Acesso não autorizado, necessário fazer login"
          },
          "404": {
            "description": "Usuário não encontrado."
          },
          "500": {
            "description": "Server ou banco fora do ar."
          }
        }
      },
      "post": {
        "tags": [
          "Usuario"
        ],
        "description": "Endpoint para criar um novo usuário.",
        "produces": [
          "application/json"
        ],
        "parameters": [],
        "responses": {
          "200": {
            "description": "Usuário criado."
          },
          "401": {
            "description": " Acesso não autorizado, necessário fazer login"
          },
          "500": {
            "description": "Server ou banco fora do ar."
          }
        }
      }
    },
    "/usuario/{id}": {
      "get": {
        "tags": [
          "usuario"
        ],
        "description": "Endpoint que encontra um usuário através de seu id.",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string",
            "description": "ID do usuário"
          }
        ],
        "responses": {
          "200": {
            "schema": {
              "$ref": "#/definitions/Projeto"
            },
            "description": "Usuário existe e é encontrado."
          },
          "401": {
            "description": " Acesso não autorizado, necessário fazer login"
          },
          "404": {
            "description": "Usuário não encontrado"
          },
          "500": {
            "description": "Server ou banco fora do ar."
          }
        }
      },
      "put": {
        "tags": [
          "Usuario"
        ],
        "description": "Endpoint para encontrar e atualizar um usuário através do id.",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string",
            "description": "ID do usuário"
          }
        ],
        "responses": {
          "200": {
            "schema": {
              "$ref": "#/definitions/Usuario"
            },
            "description": "Usuário foi editado."
          },
          "401": {
            "description": " Acesso não autorizado, necessário fazer login"
          },
          "404": {
            "description": "Usuário não encontrado"
          },
          "500": {
            "description": "Server ou banco fora do ar."
          }
        }
      },
      "delete": {
        "tags": [
          "Usuario"
        ],
        "description": "Endpoint para encontrar e excluir um usuário através do id.",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string",
            "description": "ID do usuário"
          }
        ],
        "responses": {
          "401": {
            "description": " Acesso não autorizado, necessário fazer login"
          },
          "404": {
            "description": "Usuário não encontrado"
          },
          "500": {
            "description": "Server ou banco fora do ar."
          }
        }
      }
    }
  },
  "definitions": {
    "Projeto": {
      "type": "object",
      "properties": {
        "_id": {
          "type": "string",
          "example": "60942851c1af0e0c8c02ce0f"
        },
        "nome": {
          "type": "string",
          "example": "Projeto1"
        },
        "desc": {
          "type": "string",
          "example": "Primeiro Projeto"
        },
        "info": {
          "type": "string",
          "example": "Olá mundo"
        },
        "usuario": {
          "type": "string",
          "example": "Duds12345"
        },
        "_v": {
          "type": "number",
          "example": 0
        },
        "id": {
          "type": "string",
          "example": "60942851c1af0e0c8c02ce0f"
        }
      }
    },
    "Postagem": {
      "type": "object",
      "properties": {
        "_id": {
          "type": "string",
          "example": "60944a780b4bf75c240f5865"
        },
        "post": {
          "type": "string",
          "example": "Manda post"
        },
        "_v": {
          "type": "number",
          "example": 0
        }
      }
    },
    "Usuario": {
      "type": "object",
      "properties": {
        "_id": {
          "type": "string",
          "example": "60941f25c1af0e0c8c02cdff"
        },
        "nome": {
          "type": "string",
          "example": "Perfil1"
        },
        "bio": {
          "type": "string",
          "example": "Gosto de Macarrão"
        },
        "areaAt": {
          "type": "string",
          "example": "Computação"
        },
        "_v": {
          "type": "number",
          "example": 0
        }
      }
    },
    "SignUp": {
      "type": "object",
      "properties": {
        "success": {
          "type": "boolean",
          "example": true
        },
        "status": {
          "type": "string",
          "example": "Registration Successful!"
        }
      }
    },
    "SignUpErro": {
      "type": "object",
      "properties": {
        "err": {
          "type": "object",
          "properties": {
            "name": {
              "type": "string",
              "example": "UserExistsError"
            },
            "message": {
              "type": "string",
              "example": "A user with the given username is already registered"
            }
          }
        }
      }
    },
    "Login": {
      "type": "object",
      "properties": {
        "username": {
          "type": "string",
          "example": "any"
        },
        "password": {
          "type": "string",
          "example": "any"
        }
      }
    },
    "LoginResposta": {
      "type": "object",
      "properties": {
        "token": {
          "type": "string",
          "example": "<Bearer Token>"
        },
        "username": {
          "type": "string",
          "example": "any"
        }
      }
    },
    "Logout": {
      "type": "object",
      "properties": {
        "token": {
          "type": "string",
          "example": "<Bearer Token>"
        }
      }
    }
  },
  "security": [
    {
      "bearerAuth": []
    }
  ]
}